Changed the test file and format changes
